Add LoadLoginToken to docker command object
