+++ pkg/docker/command.go
@@ -2,4 +2,11 @@ package docker
 
 type Command interface {
 	Push(string) error
+	LoadLoginToken(string) (string, error)
+}
+
+type CredentialHelperInput struct {
+	Username  string
+	Secret    string
+	ServerURL string
 }
