+++ pkg/docker/docker_command.go
@@ -1,9 +1,15 @@
 package docker
 
 import (
+	"encoding/json"
+	"fmt"
+	"io"
 	"os"
 	"os/exec"
 	"strings"
+
+	"github.com/docker/cli/cli/config"
+	"github.com/docker/cli/cli/config/configfile"
 
 	"github.com/replicate/cog/pkg/util/console"
 )
@@ -18,6 +24,15 @@ func (c *DockerCommand) Push(image string) error {
 	return c.exec("push", image)
 }
 
+func (c *DockerCommand) LoadLoginToken(registryHost string) (string, error) {
+	conf := config.LoadDefaultConfigFile(os.Stderr)
+	credsStore := conf.CredentialsStore
+	if credsStore == "" {
+		return loadAuthFromConfig(conf, registryHost)
+	}
+	return loadAuthFromCredentialsStore(credsStore, registryHost)
+}
+
 func (c *DockerCommand) exec(name string, args ...string) error {
 	cmdArgs := []string{name}
 	cmdArgs = append(cmdArgs, args...)
@@ -27,4 +42,51 @@ func (c *DockerCommand) exec(name string, args ...string) error {
 
 	console.Debug("$ " + strings.Join(cmd.Args, " "))
 	return cmd.Run()
+}
+
+func loadAuthFromConfig(conf *configfile.ConfigFile, registryHost string) (string, error) {
+	return conf.AuthConfigs[registryHost].Password, nil
+}
+
+func loadAuthFromCredentialsStore(credsStore string, registryHost string) (string, error) {
+	var out strings.Builder
+	binary := DockerCredentialBinary(credsStore)
+	cmd := exec.Command(binary, "get")
+	cmd.Env = os.Environ()
+	cmd.Stdout = &out
+	cmd.Stderr = &out
+	stdin, err := cmd.StdinPipe()
+	if err != nil {
+		return "", err
+	}
+	defer stdin.Close()
+	console.Debug("$ " + strings.Join(cmd.Args, " "))
+	err = cmd.Start()
+	if err != nil {
+		return "", err
+	}
+	_, err = io.WriteString(stdin, registryHost)
+	if err != nil {
+		return "", err
+	}
+	err = stdin.Close()
+	if err != nil {
+		return "", err
+	}
+	err = cmd.Wait()
+	if err != nil {
+		return "", fmt.Errorf("exec wait error: %w", err)
+	}
+
+	var config CredentialHelperInput
+	err = json.Unmarshal([]byte(out.String()), &config)
+	if err != nil {
+		return "", err
+	}
+
+	return config.Secret, nil
+}
+
+func DockerCredentialBinary(credsStore string) string {
+	return "docker-credential-" + credsStore
 }
