+++ pkg/docker/login.go
@@ -3,7 +3,6 @@ package docker
 import (
 	"encoding/json"
 	"fmt"
-	"io"
 	"os"
 	"os/exec"
 	"strings"
@@ -15,12 +14,6 @@ import (
 	"github.com/replicate/cog/pkg/util/console"
 )
 
-type credentialHelperInput struct {
-	Username  string
-	Secret    string
-	ServerURL string
-}
-
 func SaveLoginToken(registryHost string, username string, token string) error {
 	conf := config.LoadDefaultConfigFile(os.Stderr)
 	credsStore := conf.CredentialsStore
@@ -28,19 +21,6 @@ func SaveLoginToken(registryHost string, username string, token string) error {
 		return saveAuthToConfig(conf, registryHost, username, token)
 	}
 	return saveAuthToCredentialsStore(credsStore, registryHost, username, token)
-}
-
-func LoadLoginToken(registryHost string) (string, error) {
-	conf := config.LoadDefaultConfigFile(os.Stderr)
-	credsStore := conf.CredentialsStore
-	if credsStore == "" {
-		return loadAuthFromConfig(conf, registryHost)
-	}
-	return loadAuthFromCredentialsStore(credsStore, registryHost)
-}
-
-func dockerCredentialBinary(credsStore string) string {
-	return "docker-credential-" + credsStore
 }
 
 func saveAuthToConfig(conf *configfile.ConfigFile, registryHost string, username string, token string) error {
@@ -56,8 +36,8 @@ func saveAuthToConfig(conf *configfile.ConfigFile, registryHost string, username
 }
 
 func saveAuthToCredentialsStore(credsStore string, registryHost string, username string, token string) error {
-	binary := dockerCredentialBinary(credsStore)
-	input := credentialHelperInput{
+	binary := DockerCredentialBinary(credsStore)
+	input := CredentialHelperInput{
 		Username:  username,
 		Secret:    token,
 		ServerURL: registryHost,
@@ -83,47 +63,4 @@ func saveAuthToCredentialsStore(credsStore string, registryHost string, username
 		return fmt.Errorf("Failed to run %s: %w", binary, err)
 	}
 	return nil
-}
-
-func loadAuthFromConfig(conf *configfile.ConfigFile, registryHost string) (string, error) {
-	return conf.AuthConfigs[registryHost].Password, nil
-}
-
-func loadAuthFromCredentialsStore(credsStore string, registryHost string) (string, error) {
-	var out strings.Builder
-	binary := dockerCredentialBinary(credsStore)
-	cmd := exec.Command(binary, "get")
-	cmd.Env = os.Environ()
-	cmd.Stdout = &out
-	cmd.Stderr = &out
-	stdin, err := cmd.StdinPipe()
-	if err != nil {
-		return "", err
-	}
-	defer stdin.Close()
-	console.Debug("$ " + strings.Join(cmd.Args, " "))
-	err = cmd.Start()
-	if err != nil {
-		return "", err
-	}
-	_, err = io.WriteString(stdin, registryHost)
-	if err != nil {
-		return "", err
-	}
-	err = stdin.Close()
-	if err != nil {
-		return "", err
-	}
-	err = cmd.Wait()
-	if err != nil {
-		return "", fmt.Errorf("exec wait error: %w", err)
-	}
-
-	var config credentialHelperInput
-	err = json.Unmarshal([]byte(out.String()), &config)
-	if err != nil {
-		return "", err
-	}
-
-	return config.Secret, nil
 }
