+++ pkg/docker/push.go
@@ -2,7 +2,7 @@ package docker
 
 func Push(image string, fast bool, projectDir string, command Command) error {
 	if fast {
-		return FastPush(image, projectDir)
+		return FastPush(image, projectDir, command)
 	}
 	return StandardPush(image, command)
 }
